DocAPI

+++ src/controller/products.controller.bd.js
@@ -19,17 +19,17 @@ const getProductsBd = async (req, res) => {
 
 const addProductBd = async (req, res, next) => {
   const product = req.body;
-  if (!product.title) {
-    return next(CustomError.createError({ code: 401, msg: invalidParamsProduct(product), typeError: ERROR_FROM_SERVER }));
+  if (req.user.role === 'premium') {
+    product.owner = req.user.email;
+    const newproduct = await ProductRepository.add(product);
+    return res.json(newproduct);
   }
-  const newproduct = await ProductRepository.add(product);
-  if (newproduct) {
-    res.json(newproduct);
-  } else {
-    res.json(newproduct);
+  if (!product.owner) {
+    const newproduct = await ProductRepository.add(product);
+    return res.json(newproduct);
   }
-};
 
+};
 
 const getProductIdBd = async (req, res) => {
   const id = req.params.pid;
@@ -54,12 +54,25 @@ const UpdateProductBd = async (req, res) => {
 
 const deleteProductBd = async (req, res) => {
   const id = req.params.pid;
-  const deleteproduct = await BdProductManager.DeleteProductId(id);
-  if (deleteproduct) {
-    res.json(deleteproduct);
-  } else {
-    res.json(deleteproduct);
+  const productExist = await BdProductManager.getProductId(id);
+  if (!productExist) {
+    return res.json({ msg: 'Producto Inexistente' });
   }
+  if (req.user.role === 'admin') {
+    const deleteproduct = await BdProductManager.DeleteProductId(id);
+    return res.json({ msg: 'Producto Eliminado' });
+  }
+  if (req.user.role === 'premium') {
+    if (req.user.email == productExist.owner) {
+      const deleteproduct = await BdProductManager.DeleteProductId(id);
+      return res.json({ msg: 'Producto Eliminado' });
+    } else {
+      return res.json({ msg: 'No tenes permisos para eliminar el producto' });
+    }
+  } else {
+    return res.json({ msg: 'No tenes permisos para eliminar el producto' });
+  }
+
 };
 
 module.exports = {
